change tx fee type

+++ types/tx_type_txt.go
@@ -162,7 +162,7 @@ func (tx *TokenTransaction) IllegalGasLimitOrGasPrice(hascode bool) bool {
 		if common.IsLKC(tx.data.TokenAddress) {
 			gasFee = CalNewAmountGas(tx.Value())
 		} else {
-			gasFee = MinGasLimit
+			gasFee = uint64(MinGasLimit)
 		}
 	}
